- Fixed running the tests on Windows. - Fixed comparison bug in the main tool.
A checkout of the text files on Windows resulted in a CR/LF conversion being done. That wreaked havoc with
the MD5 and SHA* tests. I've made the tests depend on the OS to see which results are expected.

There was a bug in the main tool that forgot to actually compare the two sets of results (in the file and in the
filesystem) the other way around as well. Fixed.